Move system intake banners.
