+++ src/views/Home/index.tsx
@@ -12,9 +12,9 @@ import { useFlags } from 'contexts/flagContext';
 import { AppState } from 'reducers/rootReducer';
 import { BusinessCaseModel } from 'types/businessCase';
 import { Flags } from 'types/flags';
-import { fetchBusinessCases, fetchSystemIntakes } from 'types/routines';
-import { SystemIntakeForm } from 'types/systemIntake';
+import { fetchBusinessCases } from 'types/routines';
 
+import SystemIntakeBanners from './SystemIntakeBanners';
 import WelcomeText from './WelcomeText';
 
 import './index.scss';
@@ -25,75 +25,15 @@ const Home = ({ history }: HomeProps) => {
   const { authState } = useOktaAuth();
   const dispatch = useDispatch();
   const location = useLocation<any>();
-  const systemIntakes = useSelector(
-    (state: AppState) => state.systemIntakes.systemIntakes
-  );
   const businessCases = useSelector(
     (state: AppState) => state.businessCases.businessCases
   );
 
   useEffect(() => {
     if (authState.isAuthenticated) {
-      dispatch(fetchSystemIntakes());
       dispatch(fetchBusinessCases());
     }
   }, [dispatch, authState.isAuthenticated]);
-
-  const getSystemIntakeBanners = (flags: Flags) => {
-    const rootPath = flags.taskListLite ? '/governance-task-list' : '/system';
-    return systemIntakes.map((intake: SystemIntakeForm) => {
-      switch (intake.status) {
-        case 'INTAKE_DRAFT':
-          return (
-            <ActionBanner
-              key={intake.id}
-              title={
-                intake.requestName
-                  ? `${intake.requestName}: Intake Request`
-                  : 'Intake Request'
-              }
-              helpfulText="Your Intake Request is incomplete, please submit it when you are ready so that we can move you to the next phase"
-              onClick={() => {
-                history.push(`${rootPath}/${intake.id}`);
-              }}
-              label="Go to Intake Request"
-              data-intakeid={intake.id}
-            />
-          );
-        case 'INTAKE_SUBMITTED':
-          if (intake.businessCaseId !== null) {
-            return null;
-          }
-          return (
-            <ActionBanner
-              key={intake.id}
-              title={
-                intake.requestName
-                  ? `${intake.requestName}: Business Case`
-                  : 'Business Case'
-              }
-              helpfulText="Your intake form has been submitted. The admin team will be in touch with you to fill out a Business Case"
-              onClick={() => {
-                history.push({
-                  pathname: flags.taskListLite
-                    ? `/governance-task-list/${intake.id}`
-                    : '/business/new/general-request-info',
-                  ...((!flags.taskListLite && {
-                    state: {
-                      systemIntakeId: intake.id
-                    }
-                  }) as {})
-                });
-              }}
-              label="Start my Business Case"
-              data-intakeid={intake.id}
-            />
-          );
-        default:
-          return null;
-      }
-    });
-  };
 
   const getBusinessCaseBanners = (flags: Flags) => {
     return businessCases.map((busCase: BusinessCaseModel) => {
@@ -153,7 +93,7 @@ const Home = ({ history }: HomeProps) => {
               </p>
             </div>
           )}
-          {getSystemIntakeBanners(flags)}
+          <SystemIntakeBanners />
           {getBusinessCaseBanners(flags)}
         </div>
         <WelcomeText />
